+++ src/components/Header.vue
@@ -1,413 +1,667 @@
 <template>
-  <header class="user-header" :class="{ fixed: true }">
-    <div class="header-inner">
-      <!-- 좌측: 로고 -->
-      <div class="logo" @click="goHome">요리한수</div>
-      <!-- 중간: 메뉴바 -->
-      <nav class="menu-bar" v-if="!isMobile">
-        <a v-for="item in menuItems" :key="item.text" :href="item.route" class="menu-item" @mouseover="hoverMenu = item.text" @mouseleave="hoverMenu = ''" :class="{ active: hoverMenu === item.text }">
-          {{ item.text }}
-        </a>
-      </nav>
-      <!-- 우측: 로그인/로그아웃 버튼, 알림 버튼, 모바일 창 크기는 메뉴바 버튼 -->
-      <div class="header-actions">
-        <!-- 모바일 창 크기일 때 -->
-        <template v-if="isMobile">
-          <!-- 로그인 상태 -->
-          <template v-if="isLoggedIn">
-            <div class="notification-wrapper">
-              <button class="notification-btn" @click="toggleNotification">
-                <span class="icon-bell">
-                  <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24" viewBox="0 0 24 24" fill="none" stroke="#495057" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M18 8a6 6 0 0 0-12 0c0 7-3 9-3 9h18s-3-2-3-9"/><path d="M13.73 21a2 2 0 0 1-3.46 0"/></svg>
-                </span>
-                <span v-if="unreadCount && unreadCount > 0" class="badge">{{ unreadCount }}</span>
-              </button>
+  <v-app-bar
+    app
+    elevation="0"
+    color="white"
+    class="header-container"
+    fixed
+    prominent
+  >
+    <!-- Desktop Layout -->
+    <div class="d-none d-md-flex header-desktop">
+      <!-- Logo -->
+      <router-link to="/" class="logo-link">
+        <h1 class="logo-text">요리한수</h1>
+      </router-link>
 
-            </div>
-            <div class="cart-wrapper">
-              <button class="cart-btn" @click="goToCart">
-                <span class="icon-cart">
-                  <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24" viewBox="0 0 24 24" fill="none" stroke="#495057" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="9" cy="21" r="1"/><circle cx="20" cy="21" r="1"/><path d="M1 1h4l2.68 13.39a2 2 0 0 0 2 1.61h9.72a2 2 0 0 0 2-1.61L23 6H6"/></svg>
-                </span>
-                <span v-if="cartStore.cartCount > 0" class="cart-badge">{{ cartStore.cartCount }}</span>
-              </button>
-            </div>
-            <button class="logout-btn" @click="logout">로그아웃</button>
-          </template>
-          <!-- 로그아웃 상태 -->
-          <template v-else>
-            <button class="login-btn" @click="login">로그인</button>
-          </template>
-          <button class="menu-icon" @click="toggleMobileMenu">
-            <svg xmlns="http://www.w3.org/2000/svg" height="32" width="32" viewBox="0 0 24 24" fill="none" stroke="#FF6B35" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"><line x1="3" y1="12" x2="21" y2="12"/><line x1="3" y1="6" x2="21" y2="6"/><line x1="3" y1="18" x2="21" y2="18"/></svg>
-          </button>
-        </template>
-        <!-- 데스크탑 창 크기일 때때 -->
-        <template v-else>
-          <!-- 로그아웃 상태 -->
-          <template v-if="!isLoggedIn">
-            <button class="login-btn" @click="login">로그인</button>
-          </template>
-          <!-- 로그인 상태 -->
-          <template v-else>
-            <div class="notification-wrapper">
-              <button class="notification-btn" @click="toggleNotification">
-                <span class="icon-bell">
-                  <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24" viewBox="0 0 24 24" fill="none" stroke="#495057" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M18 8a6 6 0 0 0-12 0c0 7-3 9-3 9h18s-3-2-3-9"/><path d="M13.73 21a2 2 0 0 1-3.46 0"/></svg>
-                </span>
-                <span v-if="unreadCount && unreadCount > 0" class="badge">{{ unreadCount }}</span>
-              </button>
+      <!-- Navigation Menu -->
+      <div class="nav-menu">
+        <router-link to="/notice" class="nav-link">공지사항</router-link>
+        <router-link to="/recipes" class="nav-link">레시피 공유 게시글</router-link>
+        <router-link to="/lectures" class="nav-link">판매중인 강의</router-link>
+        <router-link to="/chat" class="nav-link">1:1채팅</router-link>
+        <router-link to="/mypage" class="nav-link">마이페이지</router-link>
+      </div>
 
-            </div>
-            <div class="cart-wrapper">
-              <button class="cart-btn" @click="goToCart">
-                <span class="icon-cart">
-                  <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24" viewBox="0 0 24 24" fill="none" stroke="#495057" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="9" cy="21" r="1"/><circle cx="20" cy="21" r="1"/><path d="M1 1h4l2.68 13.39a2 2 0 0 0 2 1.61h9.72a2 2 0 0 0 2-1.61L23 6H6"/></svg>
-                </span>
-                <span v-if="cartStore.cartCount > 0" class="cart-badge">{{ cartStore.cartCount }}</span>
-              </button>
-            </div>
-            <span class="welcome">{{ nickname }}님 반갑습니다!</span>
-            <button class="logout-btn" @click="logout">로그아웃</button>
-          </template>
-        </template>
+      <!-- Right Section -->
+      <div class="header-right">
+        <!-- Logged Out State -->
+        <div v-if="!isLoggedIn" class="login-section">
+          <v-btn
+            color="primary"
+            variant="flat"
+            rounded="lg"
+            @click="goToLogin"
+            class="login-btn"
+          >
+            로그인
+          </v-btn>
+        </div>
+
+        <!-- Logged In State -->
+        <div v-else class="user-section">
+          <v-avatar size="32" class="profile-avatar">
+            <v-img 
+              v-if="profileInfo.profileImageUrl" 
+              :src="profileInfo.profileImageUrl" 
+              alt="프로필 이미지"
+            ></v-img>
+            <v-icon v-else size="20" color="grey">mdi-account</v-icon>
+          </v-avatar>
+          
+          <!-- 알림 버튼 -->
+          <v-btn
+            icon
+            variant="text"
+            class="notification-btn"
+            @click="goToNotifications"
+          >
+            <v-badge
+              :content="unreadCount"
+              :model-value="unreadCount > 0"
+              color="error"
+              location="top end"
+            >
+              <v-icon>mdi-bell</v-icon>
+            </v-badge>
+          </v-btn>
+
+          <!-- 장바구니 버튼 -->
+          <v-btn
+            icon
+            variant="text"
+            class="cart-btn"
+            @click="goToCart"
+          >
+            <v-badge
+              :content="cartCount"
+              :model-value="cartCount > 0"
+              color="warning"
+              location="top end"
+            >
+              <v-icon>mdi-cart</v-icon>
+            </v-badge>
+          </v-btn>
+          
+          <span class="welcome-text">{{ userNickname }}님 환영합니다!</span>
+
+          <v-btn
+            variant="outlined"
+            color="grey"
+            rounded="lg"
+            @click="logout"
+            class="logout-btn"
+          >
+            로그아웃
+          </v-btn>
+        </div>
       </div>
     </div>
-    <transition name="slide-down">
-      <div v-if="isMobile && showMobileMenu" class="mobile-menu">
-        <nav class="mobile-menu-bar">
-          <a v-for="item in menuItems" :key="item.text" :href="item.route" class="mobile-menu-item" @click="closeMobileMenu">{{ item.text }}</a>
-          <a href="/cart" class="mobile-menu-item" @click="closeMobileMenu">장바구니 ({{ cartStore.cartCount }})</a>
-        </nav>
+
+    <!-- Mobile Layout -->
+    <div class="d-flex d-md-none header-mobile">
+      <!-- Logged Out State -->
+      <div v-if="!isLoggedIn" class="mobile-logged-out">
+        <v-btn
+          icon
+          variant="text"
+          @click="toggleMobileMenu"
+          class="hamburger-btn"
+        >
+          <v-icon>mdi-menu</v-icon>
+        </v-btn>
+        
+        <div class="mobile-logo-center">
+          <router-link to="/" class="logo-link">
+            <h1 class="logo-text">요리한수</h1>
+          </router-link>
+        </div>
+        <v-btn
+          color="primary"
+          variant="flat"
+          rounded="lg"
+          @click="goToLogin"
+          class="mobile-login-btn"
+        >
+          로그인
+        </v-btn>
       </div>
-    </transition>
-  </header>
+
+      <!-- Logged In State -->
+      <div v-else class="mobile-logged-in">
+        <v-btn
+          icon
+          variant="text"
+          @click="toggleMobileMenu"
+          class="hamburger-btn"
+        >
+          <v-icon>mdi-menu</v-icon>
+        </v-btn>
+
+        <div class="mobile-logo-center">
+          <router-link to="/" class="logo-link">
+            <h1 class="logo-text">요리한수</h1>
+          </router-link>
+        </div>
+
+        <div class="mobile-user-section">
+          <!-- 모바일 알림 버튼 -->
+          <v-btn
+            icon
+            variant="text"
+            size="small"
+            class="mobile-notification-btn"
+            @click="goToNotifications"
+          >
+            <v-badge
+              :content="unreadCount"
+              :model-value="unreadCount > 0"
+              color="error"
+              location="top end"
+            >
+              <v-icon size="20">mdi-bell</v-icon>
+            </v-badge>
+          </v-btn>
+
+          <!-- 모바일 장바구니 버튼 -->
+          <v-btn
+            icon
+            variant="text"
+            size="small"
+            class="mobile-cart-btn"
+            @click="goToCart"
+          >
+            <v-badge
+              :content="cartCount"
+              :model-value="cartCount > 0"
+              color="warning"
+              location="top end"
+            >
+              <v-icon size="20">mdi-cart</v-icon>
+            </v-badge>
+          </v-btn>
+
+          <v-avatar size="32" class="profile-avatar">
+            <v-img 
+              v-if="userProfileImage && userProfileImage !== '/default-avatar.png'" 
+              :src="userProfileImage" 
+              alt="프로필 이미지"
+            ></v-img>
+            <v-icon v-else size="20" color="grey">mdi-account</v-icon>
+          </v-avatar>
+          <v-btn
+            variant="outlined"
+            color="grey"
+            rounded="lg"
+            size="small"
+            @click="logout"
+            class="mobile-logout-btn"
+          >
+            로그아웃
+          </v-btn>
+        </div>
+      </div>
+    </div>
+
+  </v-app-bar>
+
+  <!-- Mobile Navigation Drawer - v-app-bar 외부로 이동 -->
+  <v-navigation-drawer
+    v-model="mobileMenuOpen"
+    location="left"
+    temporary
+    class="mobile-nav-drawer"
+    overlay
+  >
+    <v-list>
+      <v-list-item
+        v-for="item in mobileMenuItems"
+        :key="item.path"
+        :to="item.path"
+        @click="mobileMenuOpen = false"
+      >
+        <v-list-item-title>{{ item.title }}</v-list-item-title>
+      </v-list-item>
+      <!-- 모바일 메뉴에 장바구니 추가 -->
+      <v-list-item
+        to="/cart"
+        @click="mobileMenuOpen = false"
+        class="cart-menu-item"
+      >
+        <v-list-item-title>
+          장바구니
+          <v-badge
+            :content="cartCount"
+            :model-value="cartCount > 0"
+            color="warning"
+            location="top end"
+            class="cart-menu-badge"
+          >
+          </v-badge>
+        </v-list-item-title>
+      </v-list-item>
+    </v-list>
+  </v-navigation-drawer>
 </template>
 
 <script setup>
-import { ref, computed, watch, onMounted, onUnmounted } from 'vue'
+import { ref, computed, watch, onUnmounted, onMounted } from 'vue'
 import { useRouter } from 'vue-router'
+import { useAuthStore } from '@/store/auth/auth'
 import { useCartStore } from '@/store/cart/cart.js'
 import { useNotificationStore } from '@/store/notification/notification.js'
-import { useNotifications } from '@/composables/useNotifications.js'
 
-const router = useRouter()
-const cartStore = useCartStore()
-const notificationStore = useNotificationStore()
-const { isConnected } = useNotifications()
+const router = useRouter();
+const authStore = useAuthStore();
+const cartStore = useCartStore();
+const notificationStore = useNotificationStore();
 
-// unreadCount를 computed로 참조
-const unreadCount = computed(() => notificationStore.unreadCount)
+// Reactive data
+const mobileMenuOpen = ref(false);
+const mobileMenuItems = ref([
+  { title: '공지사항', path: '/notice' },
+  { title: '레시피 공유 게시글', path: '/recipes' },
+  { title: '판매중인 강의', path: '/lectures' },
+  { title: '1:1채팅', path: '/chat' },
+  { title: '마이페이지', path: '/mypage' }
+]);
 
-// unreadCount 변화 감지
-watch(unreadCount, (newValue, oldValue) => {
-  
-}, { immediate: true })
+// 화면 크기 감지 함수
+const handleResize = () => {
+  // 데스크톱 크기(960px 이상)로 변경되면 모바일 메뉴 닫기
+  if (window.innerWidth >= 960 && mobileMenuOpen.value) {
+    mobileMenuOpen.value = false;
+  }
+}
 
-const isLoggedIn = ref(true) // 임시: 실제 로그인 상태와 연동 필요 (테스트용으로 true 설정)
-const nickname = ref('김요리') // 임시: 실제 닉네임 연동 필요
-const hoverMenu = ref('')
-const showNotification = ref(false)
-const isMobile = ref(false)
-const showMobileMenu = ref(false)
+// Reactive data for profile
+const profileData = ref({
+  nickname: '',
+  profileImageUrl: ''
+})
 
-const menuItems = [
-  { text: '공지사항', route: '/notice' },
-  { text: '레시피 공유 게시글', route: '/recipes' },
-  { text: '판매중인 강의', route: '/lectures' },
-  { text: '1:1채팅', route: '/chat' },
-  { text: '마이페이지', route: '/mypage' }
-]
+// Computed properties
+const isLoggedIn = computed(() => {
+  return authStore.getIsAuthenticated;
+})
 
-function goHome() {
-  router.push('/')
+// 프로필 정보 가져오기
+const fetchProfileInfo = async () => {
+  if (isLoggedIn.value && authStore.accessToken) {
+    try {
+      const profileInfo = await authStore.fetchProfileInfo();
+      if (profileInfo) {
+        profileData.value = {
+          nickname: profileInfo.nickname || '사용자',
+          profileImageUrl: profileInfo.profileImageUrl || ''
+        };
+      } else {
+        // 프로필 정보가 없는 경우 기본값 설정
+        profileData.value = {
+          nickname: '사용자',
+          profileImageUrl: ''
+        };
+      }
+    } catch (error) {
+      console.error('프로필 정보 가져오기 실패:', error);
+      // 에러 발생 시 기본값 설정
+      profileData.value = {
+        nickname: '사용자',
+        profileImageUrl: ''
+      };
+    }
+  }
 }
-function login() {
-  router.push('/login')
-  closeMobileMenu()
-}
-function logout() {
-  isLoggedIn.value = false
-  closeMobileMenu()
-}
-function toggleNotification() {
-  // 알림 페이지로 이동
-  router.push('/notifications')
-}
-function handleResize() {
-  isMobile.value = window.innerWidth <= 900
-  if (!isMobile.value) showMobileMenu.value = false
-}
-function toggleMobileMenu() {
-  showMobileMenu.value = !showMobileMenu.value
-}
-function closeMobileMenu() {
-  showMobileMenu.value = false
-}
-function goToCart() {
-  router.push('/cart')
-  closeMobileMenu()
-}
+
+// 모바일 메뉴 상태 감시하여 스크롤 제어
+watch(mobileMenuOpen, (isOpen) => {
+  if (isOpen) {
+    // 메뉴가 열릴 때 스크롤 방지
+    document.body.style.overflow = 'hidden';
+    document.body.style.position = 'fixed';
+    document.body.style.width = '100%';
+  } else {
+    // 메뉴가 닫힐 때 스크롤 복원
+    document.body.style.overflow = '';
+    document.body.style.position = '';
+    document.body.style.width = '';
+  }
+})
+
+// 로그인 상태 변경 감시하여 프로필 정보 가져오기
+watch(isLoggedIn, async (newValue) => {
+  if (newValue) {
+    await fetchProfileInfo();
+  } else {
+    // 로그아웃 시 프로필 정보 초기화
+    profileData.value = {
+      nickname: '',
+      profileImageUrl: ''
+    };
+  }
+})
+
+// 컴포넌트 마운트 시 리사이즈 이벤트 리스너 추가 및 프로필 정보 가져오기
 onMounted(async () => {
-  handleResize()
-  window.addEventListener('resize', handleResize)
+  window.addEventListener('resize', handleResize);
   
-  // 실시간 알림 연결은 useNotifications에서 자동 처리
-  console.log('Header 컴포넌트 마운트 완료')
+  // 로그인된 상태라면 프로필 정보 가져오기
+  if (isLoggedIn.value) {
+    await fetchProfileInfo();
+  }
 })
+
+// 컴포넌트 언마운트 시 이벤트 리스너 제거 및 스크롤 복원
 onUnmounted(() => {
-  window.removeEventListener('resize', handleResize)
+  window.removeEventListener('resize', handleResize);
+  document.body.style.overflow = '';
+  document.body.style.position = '';
+  document.body.style.width = '';
 })
+
+const userNickname = computed(() => {
+  return profileData.value.nickname || '사용자';
+})
+
+const userProfileImage = computed(() => {
+  return profileData.value.profileImageUrl || '';
+})
+
+const profileInfo = computed(() => {
+  return profileData.value;
+})
+
+// 장바구니 개수
+const cartCount = computed(() => {
+  return cartStore.cartCount;
+})
+
+// 읽지 않은 알림 개수
+const unreadCount = computed(() => {
+  return notificationStore.unreadCount;
+})
+
+// Methods
+const goToLogin = () => {
+  router.push('/login');
+}
+
+const logout = async () => {
+  try {
+    await authStore.logout();
+    // 루트 페이지로 이동 (로그아웃 후 비로그인 상태이므로 landing 페이지로 리다이렉트됨)
+    router.push('/');
+  } catch (error) {
+    console.error('로그아웃 실패:', error);
+  }
+}
+
+const goToNotifications = () => {
+  router.push('/notifications');
+}
+
+const goToCart = () => {
+  router.push('/cart');
+}
+
+const toggleMobileMenu = () => {
+  mobileMenuOpen.value = !mobileMenuOpen.value
+}
 </script>
 
 <style scoped>
-.user-header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  background: var(--color-white);
-  box-shadow: 0 2px 8px rgba(0,0,0,0.03);
+.header-container {
+  border-bottom: 1px solid #e0e0e0;
   z-index: 1000;
-  font-family: 'Noto Sans', sans-serif;
+  position: fixed !important;
+  top: 0 !important;
+  left: 0 !important;
+  right: 0 !important;
+  width: 100% !important;
 }
-.header-inner {
-  max-width: 1400px;
-  margin: 0 auto;
-  display: flex;
-  align-items: center;
+
+/* 고정된 헤더 아래 콘텐츠 패딩 */
+:deep(.v-main) {
+  padding-top: 80px !important; /* prominent 모드에서 v-app-bar의 높이 */
+}
+
+@media (max-width: 960px) {
+  :deep(.v-main) {
+    padding-top: 72px !important; /* 모바일에서 prominent 모드 v-app-bar의 높이 */
+  }
+}
+
+/* 추가 고정 설정 */
+:deep(.v-application) {
+  padding-top: 0 !important;
+}
+
+:deep(.v-main__wrap) {
+  padding-top: 0 !important;
+}
+
+/* 전역 스타일로 헤더 고정 강화 */
+:global(.v-application) {
+  padding-top: 0 !important;
+}
+
+:global(.v-main) {
+  padding-top: 80px !important;
+}
+
+@media (max-width: 960px) {
+  :global(.v-main) {
+    padding-top: 72px !important;
+  }
+}
+
+/* Desktop Styles */
+.header-desktop {
+  width: 100%;
   justify-content: space-between;
-  height: 64px;
-  padding: 0 32px;
+  align-items: center;
+  padding: 0 24px;
 }
-.logo {
-  color: var(--color-primary);
-  font-size: 2rem;
-  font-weight: bold;
-  cursor: pointer;
-  user-select: none;
-  text-align: left;
-}
-.menu-bar {
-  display: flex;
-  gap: 40px;
-}
-.menu-item {
-  color: var(--color-text);
-  font-size: 1.1rem;
+
+.logo-link {
   text-decoration: none;
-  padding: 8px 0;
-  position: relative;
-  transition: color 0.2s;
 }
-.menu-item.active,
-.menu-item:hover {
+
+.logo-text {
+  color: var(--color-primary);
+  font-size: 24px;
+  font-weight: bold;
+  margin: 0;
+}
+
+.nav-menu {
+  display: flex;
+  gap: 32px;
+  align-items: center;
+}
+
+.nav-link {
+  color: var(--color-text);
+  text-decoration: none;
+  font-size: 14px;
+  transition: color 0.3s ease;
+}
+
+.nav-link:hover {
   color: var(--color-primary);
 }
-.menu-item.active::after,
-.menu-item:hover::after {
-  content: '';
-  display: block;
-  height: 3px;
-  background: var(--color-primary);
-  border-radius: 2px;
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: -4px;
+
+.nav-link.router-link-active {
+  color: var(--color-primary);
+  font-weight: 500;
 }
-.header-actions {
+
+.header-right {
   display: flex;
   align-items: center;
   gap: 16px;
 }
-.login-btn, .logout-btn {
-  background: var(--color-primary);
-  color: #fff;
-  border: none;
-  border-radius: 8px;
-  padding: 8px 20px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background 0.2s;
-}
-.login-btn:hover, .logout-btn:hover {
-  background: #ff884d;
-}
-.welcome {
-  color: var(--color-text);
-  font-size: 1rem;
-  margin: 0 8px;
-}
-.notification-wrapper {
-  position: relative;
-}
-.notification-btn {
-  background: none;
-  border: none;
-  position: relative;
-  font-size: 1.5rem;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  padding: 0;
-}
-.badge {
-  position: absolute;
-  top: -6px;
-  right: -8px;
-  background: #e53935;
-  color: #fff;
-  border-radius: 50%;
-  font-size: 0.8rem;
-  padding: 2px 6px;
-  min-width: 20px;
-  text-align: center;
-  font-weight: bold;
+
+.login-btn {
+  font-weight: 500;
 }
 
-.cart-wrapper {
+.user-section {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.profile-avatar {
+  border: 2px solid var(--color-primary);
+}
+
+.notification-btn {
   position: relative;
 }
+
 .cart-btn {
-  background: none;
-  border: none;
   position: relative;
-  font-size: 1.5rem;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  padding: 0;
 }
-.cart-badge {
-  position: absolute;
-  top: -6px;
-  right: -8px;
-  background: #ff7a00;
-  color: #fff;
-  border-radius: 50%;
-  font-size: 0.8rem;
-  padding: 2px 6px;
-  min-width: 20px;
-  text-align: center;
-  font-weight: bold;
-}
-.fade-slide-enter-active, .fade-slide-leave-active {
-  transition: opacity 0.25s, transform 0.25s;
-}
-.fade-slide-enter-from, .fade-slide-leave-to {
-  opacity: 0;
-  transform: translateY(-10px);
-}
-.fade-slide-enter-to, .fade-slide-leave-from {
-  opacity: 1;
-  transform: translateY(0);
-}
-.menu-icon {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  display: flex;
-  align-items: center;
-}
-@media (max-width: 1000px) {
-  .header-inner {
-    padding: 0 12px;
-    gap: 12px;
-  }
-  .menu-bar {
-    gap: 18px;
-  }
-  .logo {
-    font-size: 1.3rem;
-  }
-}
-@media (max-width: 900px) {
-  .header-inner {
-    justify-content: center;
-    padding: 0 8px;
-    gap: 0;
-  }
-  .logo {
-    flex: none;
-    text-align: center;
-    margin: 0 auto;
-    position: absolute;
-    left: 0; right: 0;
-    top: 0; bottom: 0;
-    height: 64px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  .header-actions {
-    position: absolute;
-    right: 12px;
-    top: 16px;
-    gap: 0;
-  }
-  .menu-bar {
-    display: none;
-  }
-  .header-actions {
-    gap: 10px;
-  }
-  .logout-btn, .login-btn {
-    padding: 6px 12px;
-    font-size: 0.98rem;
-  }
-}
-.slide-down-enter-active, .slide-down-leave-active {
-  transition: max-height 0.3s cubic-bezier(.4,0,.2,1), opacity 0.3s cubic-bezier(.4,0,.2,1);
-}
-.slide-down-enter-from, .slide-down-leave-to {
-  max-height: 0;
-  opacity: 0;
-}
-.slide-down-enter-to, .slide-down-leave-from {
-  max-height: 300px;
-  opacity: 1;
-}
-.mobile-menu {
-  background: var(--color-white);
-  box-shadow: 0 4px 16px rgba(0,0,0,0.08);
-  border-radius: 0 0 16px 16px;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 64px;
-  z-index: 1500;
-  padding: 16px 0 8px 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.mobile-menu-bar {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-  width: 100%;
-  align-items: center;
-}
-.mobile-menu-item {
+
+.welcome-text {
   color: var(--color-text);
-  font-size: 1.1rem;
-  text-decoration: none;
-  padding: 8px 0;
+  font-size: 14px;
+  white-space: nowrap;
+}
+
+.logout-btn {
+  font-size: 12px;
+  font-weight: 500;
+}
+
+/* Mobile Styles */
+.header-mobile {
   width: 100%;
-  text-align: center;
-  transition: color 0.2s;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 16px;
 }
-.mobile-menu-item:active,
-.mobile-menu-item:hover {
-  color: var(--color-primary);
-}
-.mobile-user-actions {
-  margin-top: 16px;
+
+.mobile-logged-out {
+  width: 100%;
   display: flex;
-  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.mobile-logged-in {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.mobile-logo-center {
+  flex: 1;
+  display: flex;
+  justify-content: center;
+}
+
+.mobile-logo-center .logo-text {
+  font-size: 20px;
+}
+
+.mobile-user-section {
+  display: flex;
   align-items: center;
   gap: 8px;
 }
-</style> 
+
+.mobile-notification-btn,
+.mobile-cart-btn {
+  position: relative;
+}
+
+.hamburger-btn {
+  color: var(--color-text);
+  transition: color 0.3s ease;
+}
+
+.hamburger-btn:hover {
+  color: var(--color-primary);
+}
+
+.mobile-login-btn,
+.mobile-logout-btn {
+  font-size: 12px;
+  font-weight: 500;
+}
+
+.mobile-nav-drawer {
+  background-color: white;
+  z-index: 1001;
+  box-shadow: 2px 0 8px rgba(0, 0, 0, 0.1);
+}
+
+/* 모바일 메뉴 오버레이 스타일 */
+:deep(.v-overlay) {
+  z-index: 1000;
+}
+
+.mobile-nav-drawer .v-list-item {
+  border-bottom: 1px solid #f0f0f0;
+  padding: 16px 24px;
+}
+
+.mobile-nav-drawer .v-list-item:hover {
+  background-color: #f8f9fa;
+}
+
+.mobile-nav-drawer .v-list-item-title {
+  color: var(--color-text);
+  font-size: 16px;
+  font-weight: 500;
+}
+
+.mobile-nav-drawer .v-list-item.router-link-active {
+  background-color: var(--color-primary);
+  color: white;
+}
+
+.mobile-nav-drawer .v-list-item.router-link-active .v-list-item-title {
+  color: white;
+}
+
+.cart-menu-item {
+  border-bottom: 1px solid #f0f0f0;
+  padding: 16px 24px;
+}
+
+.cart-menu-item:hover {
+  background-color: #f8f9fa;
+}
+
+.cart-menu-item .v-list-item-title {
+  color: var(--color-text);
+  font-size: 16px;
+  font-weight: 500;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.cart-menu-badge {
+  margin-left: auto;
+}
+
+/* Responsive adjustments */
+@media (max-width: 600px) {
+  .header-mobile {
+    padding: 0 12px;
+  }
+  
+  .mobile-logo-center .logo-text {
+    font-size: 18px;
+  }
+  
+  .mobile-user-section {
+    gap: 6px;
+  }
+}
+</style>
