+++ src/views/MyPage/MyPage.vue
@@ -75,7 +75,7 @@
 </template>
 
 <script>
-import Header from '@/components/Header1.vue';
+import Header from '@/components/Header.vue';
 import MyRecipes from '@/components/mypage/MyRecipes.vue';
 import MyPosts from '@/components/mypage/MyPosts.vue';
 import PurchasedLectures from '@/components/mypage/PurchasedLectures.vue';
