+++ src/views/home/RecipeMainPage.vue
@@ -77,7 +77,7 @@
 </template>
 
 <script>
-import Header from '@/components/Header1.vue';
+import Header from '@/components/Header.vue';
 import Pagination from '@/components/common/Pagination.vue';
 
 const defaultThumbnail = '/src/assets/images/smu_mascort1.jpg';
