+++ src/views/payment/PaymentDetails.vue
@@ -131,7 +131,7 @@
 </template>
 
 <script>
-import Header from '@/components/Header1.vue';
+import Header from '@/components/Header.vue';
 import Footer from '@/components/Footer.vue';
 
 export default {
